rm and mv functionality added
I've added the rm and mv functionality and it works quite well, imo. I still need to test it a bit more. There are some issues with commands not running on Oak when I try them out. I'm referring to the find commands he mentioned in the pdf. They don't seem to work on the server at all.